Isolate AnonUser and UserSession classes
* Remove the AnonUser and UserSession classes into separate modules for
  easy reference.
* Move some utility function out a separate module.
